requirements.txt
discord.py[voice]
python-dotenv
flask
yt-dlp
